De-nest source files from the plugins directory
This commit is coming after cherry-picking the commits that made up
Names DB from the monorepo onto a new orphan branch. All the code was
obviously left in the nested directories from before though; so this
commit moves everything up to the root of the repository.

It also removes some latent references to other parts of the codebase,
like the JavaScript SDK in the GitLab CI file.